+++ lib/vey-reuseport/src/tcp.rs
@@ -95,7 +95,7 @@ impl TcpSocketSelector {
                 .as_mut()
         } {
             ro_data.load_pid = self.pid;
-            ro_data.load_generation = self.generation;
+            ro_data.load_generation = self.generation as u32;
         }
 
         Ok(open_object)
