+++ lib/vey-reuseport/src/udp.rs
@@ -106,7 +106,7 @@ impl UdpSocketSelector {
                 .as_mut()
         } {
             ro_data.load_pid = self.pid;
-            ro_data.load_generation = self.generation;
+            ro_data.load_generation = self.generation as u32;
         }
 
         Ok(open_object)
